requirements.txt
tensorflow-cpu
streamlit
librosa
numpy
matplotlib
scikit-learn
google-genai
pillow
python-dotenv
